+++ android-example/src/main/java/org/rewedigital/katana/android/example/fragment/SecondFragmentModule.kt
@@ -3,7 +3,7 @@ package org.rewedigital.katana.android.example.fragment
 import org.rewedigital.katana.Module
 import org.rewedigital.katana.android.example.fragment.inject.Container
 import org.rewedigital.katana.android.example.fragment.model.SecondFragmentViewModel
-import org.rewedigital.katana.androidx.viewmodel.viewModel
+import org.rewedigital.katana.androidx.viewmodel.savedstate.viewModelSavedState
 import org.rewedigital.katana.dsl.compact.factory
 import org.rewedigital.katana.dsl.get
 
@@ -15,5 +15,5 @@ val secondFragmentModule = Module {
 
     factory { Container(get(SOME_DEPENDENCY), get(FRAGMENT_DEPENDENCY2)) }
 
-    viewModel { SecondFragmentViewModel() }
+    viewModelSavedState { state -> SecondFragmentViewModel(state) }
 }